pyproject.toml
[tool.poetry]
name = "nytapi"
version = "0.1.0"
description = "nyt summarizer"
authors = ["dhpitt <david.h.pitt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pynytimes = "^0.9.0"
openai = "^0.27.4"
python-crontab = "^2.7.1"
google-api-python-client = "^2.85.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
black = "^23.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
